correct_contollers_mappings v1
