+++ webapp/src/main/java/jm/controller/CreateWorkspaceController.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
+@RequestMapping(value = "/createWorkspace/**")
 public class CreateWorkspaceController {
 
     @GetMapping(value = "/create")
