+++ webapp/src/main/java/jm/controller/MailTestController.java
@@ -3,8 +3,10 @@ package jm.controller;
 import jm.MailService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping(value = "/email/**")
 public class MailTestController {
 
     private MailService mailService;
