+++ webapp/src/main/java/jm/controller/MainController.java
@@ -26,25 +26,11 @@ public class MainController {
         return "home-page";
     }
 
-    @PostMapping(value = "/workspace/create")
-    public ModelAndView addUser(@RequestParam("name") String name, @RequestParam("usersList") String[] usersList,
-                                @RequestParam("owner") String owner, @RequestParam(value = "isPrivate", required = false) boolean isPrivate) {
-        ModelAndView modelAndView = new ModelAndView();
-        //TODO
-        modelAndView.setViewName("redirect:/workspace");
-        return modelAndView;
-    }
-
-    @GetMapping(value = "/workspace")
-    public ModelAndView workspacePage() {
-        return new ModelAndView("workspace-page");
-    }
-
+   /*
     @GetMapping(value = "/workspace_temp")
     public ModelAndView workspaceTempPage() {
         return new ModelAndView("temp/workspace-page-temp.html");
-    }
-
+*/
     @GetMapping(value = "/signin")
     public ModelAndView signInPage() {
         return new ModelAndView("signin-page");
