+++ webapp/src/main/resources/application.properties
@@ -1,8 +1,11 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/jm_message_system?characterEncoding=UTF-8&useUnicode=true&useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=123
+spring.profiles.active=default
+server.port=8080
+
+spring.datasource.url=jdbc:mysql://localhost:3306/test_db?characterEncoding=UTF-8&useUnicode=true&useSSL=false&serverTimezone=UTC
+spring.datasource.username=username
+spring.datasource.password=password
 spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
+
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 
@@ -16,3 +19,5 @@ spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 
 spring.kafka.bootstrap-servers=localhost:9092
+
+logging.config=logback-spring.xml
